+++ src/Node.h
@@ -5,24 +5,23 @@
 
 class Node
 {
-	std::vector< pair< unsigned int, short > > adjacentNodes;
+	std::vector< pair< unsigned int, float > > adjacentNodes; // toID weight
 	unsigned int ID;
 	float x, y;
 
 public:
 
 	// Constructor
-	//Node(unsigned int ID);
 	Node(unsigned int ID, float x, float y);
 
 	// Accessors
 	unsigned int getNodeID() const;
 	float getX() const;
 	float getY() const;
-	std::vector< pair< unsigned int, short > > getAdjacentNodes() const;
+	std::vector< pair< unsigned int, float > > getAdjacentNodes() const;
 
 	//Mutators
-	void setAdjacentNodes(unsigned int toID, short weight);
+	void setAdjacentNodes(unsigned int toID, float weight);
 
 	// Helper Functions
 	void Draw() const;
